+++ typescript/ui/src/ragbits/stores/HistoryStore/eventHandlers/messageHandlers.ts
@@ -9,6 +9,7 @@ import {
   MessageUsageChatResponse,
   ReferenceChatResponse,
   TextChatResponse,
+  TodoItemChatResonse,
 } from "@ragbits/api-client-react";
 import { PrimaryHandler } from "./eventHandlerRegistry";
 import { produce } from "immer";
@@ -150,4 +151,23 @@ export const handleError: PrimaryHandler<ErrorChatResponse> = (
 ) => {
   const message = draft.history[ctx.messageId];
   message.error = response.content.message;
+};
+
+export const handleTodoItem: PrimaryHandler<TodoItemChatResonse> = (
+  { content },
+  draft,
+  ctx,
+) => {
+  const message = draft.history[ctx.messageId];
+  const tasks = message.tasks ?? [];
+  const newTasks = produce(tasks, (tasksDraft) => {
+    const taskIndex = tasksDraft.findIndex((t) => t.id === content.id);
+    if (taskIndex === -1) {
+      tasksDraft.push(content);
+    } else {
+      tasksDraft[taskIndex] = content;
+    }
+  });
+
+  message.tasks = newTasks;
 };